+++ app/src/main/java/org/torproject/android/OrbotMainActivity.java
@@ -16,7 +16,7 @@ import java.util.Random;
 import java.util.StringTokenizer;
 
 import org.json.JSONArray;
-import org.torproject.android.hsutils.GrantedPermissionsAction;
+import org.torproject.android.storage.GrantedPermissionsAction;
 import org.torproject.android.service.OrbotConstants;
 import org.torproject.android.service.util.Prefs;
 import org.torproject.android.service.TorService;
